task_certification_platform/requirements.txt
# Core Dependencies
Django==4.2.7
djangorestframework==3.14.0
django-crispy-forms==2.0
crispy-bootstrap5==2023.10
Pillow==10.0.1
reportlab==4.0.7
django-filter==23.3
python-dateutil==2.8.2
django-widget-tweaks==1.5.0
python-dotenv==1.0.0

# Development
django-debug-toolbar==4.1.0
ipython==8.12.0
pytest==7.3.1
pytest-django==4.5.2
coverage==7.2.7

# Production (uncomment for production)
# gunicorn==20.1.0
# psycopg2-binary==2.9.6
# whitenoise==6.4.0

# Documentation
sphinx==7.0.1
sphinx-rtd-theme==1.2.0
